pyproject.toml
[project]
name = "backend-mpa"
version = "0.1.0"
description = ""
authors = [
    {name = "Non-list-zmlskty",email = "juanoz7u7@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask (==2.3.3)",
    "flask-cors (==4.0.0)",
    "sqlalchemy (==1.4.50)",
    "pymysql (==1.1.0)",
    "python-dotenv (==1.0.0)",
    "pyjwt (==2.8.0)",
    "passlib[bcrypt] (==1.7.4)",
    "bcrypt (==4.0.1)",
    "pyngrok (==7.0.0)",
    "email-validator (==2.0.0)",
    "cryptography (==39.0.2)",
    "pydantic (==1.10.13)"
]

[tool.poetry]
packages = [{include = "backend_mpa", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
